feat: create top section state holder
