+++ src/jvmMain/kotlin/jp/kaleidot725/adbpad/Main.kt
@@ -1,14 +1,5 @@
 import androidx.compose.animation.Crossfade
-import androidx.compose.animation.core.animateFloatAsState
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.wrapContentSize
-import androidx.compose.foundation.layout.wrapContentWidth
 import androidx.compose.material.Colors
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Surface
@@ -18,51 +9,32 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.ExperimentalComposeUiApi
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.input.pointer.PointerEventType
-import androidx.compose.ui.input.pointer.onPointerEvent
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.Window
 import androidx.compose.ui.window.WindowScope
 import androidx.compose.ui.window.WindowState
 import androidx.compose.ui.window.application
-import com.composables.icons.lucide.Camera
-import com.composables.icons.lucide.Circle
-import com.composables.icons.lucide.Lucide
-import com.composables.icons.lucide.Power
-import com.composables.icons.lucide.PowerOff
-import com.composables.icons.lucide.RefreshCcw
-import com.composables.icons.lucide.Square
-import com.composables.icons.lucide.Triangle
-import com.composables.icons.lucide.Volume1
-import com.composables.icons.lucide.Volume2
 import jp.kaleidot725.adbpad.MainCategory
-import jp.kaleidot725.adbpad.MainState
 import jp.kaleidot725.adbpad.MainStateHolder
 import jp.kaleidot725.adbpad.domain.di.domainModule
 import jp.kaleidot725.adbpad.domain.model.Dialog
 import jp.kaleidot725.adbpad.domain.model.UserColor
-import jp.kaleidot725.adbpad.domain.model.device.Device
 import jp.kaleidot725.adbpad.domain.model.language.Language
 import jp.kaleidot725.adbpad.domain.model.setting.WindowSize
 import jp.kaleidot725.adbpad.domain.model.setting.getWindowSize
 import jp.kaleidot725.adbpad.repository.di.repositoryModule
-import jp.kaleidot725.adbpad.ui.component.CommandIconButton
-import jp.kaleidot725.adbpad.ui.component.CommandIconDivider
 import jp.kaleidot725.adbpad.ui.component.NavigationRail
 import jp.kaleidot725.adbpad.ui.di.stateHolderModule
 import jp.kaleidot725.adbpad.ui.screen.CommandScreen
 import jp.kaleidot725.adbpad.ui.screen.ScreenLayout
 import jp.kaleidot725.adbpad.ui.screen.error.AdbErrorScreen
-import jp.kaleidot725.adbpad.ui.screen.menu.component.DropDownDeviceMenu
 import jp.kaleidot725.adbpad.ui.screen.screenshot.ScreenshotScreen
 import jp.kaleidot725.adbpad.ui.screen.setting.SettingScreen
 import jp.kaleidot725.adbpad.ui.screen.setting.SettingStateHolder
 import jp.kaleidot725.adbpad.ui.screen.text.TextCommandScreen
+import jp.kaleidot725.adbpad.ui.section.TopSection
 import org.koin.core.context.GlobalContext
 import org.koin.core.context.startKoin
 
@@ -116,9 +88,11 @@ fun WindowScope.App(mainStateHolder: MainStateHolder) {
         Surface {
             ScreenLayout(
                 top = {
-                    TitleBarView(
-                        state = state,
-                        onSelectDevice = mainStateHolder::selectDevice,
+                    val topStateHolder = mainStateHolder.topStateHolder
+                    val topState by topStateHolder.state.collectAsState()
+                    TopSection(
+                        state = topState,
+                        onSelectDevice = topStateHolder::selectDevice,
                         onRefresh = mainStateHolder::refresh,
                     )
                 },
@@ -247,96 +221,6 @@ fun WindowScope.App(mainStateHolder: MainStateHolder) {
                 },
                 modifier = Modifier.fillMaxSize(),
             )
-        }
-    }
-}
-
-@OptIn(ExperimentalComposeUiApi::class)
-@Composable
-private fun TitleBarView(
-    state: MainState,
-    onSelectDevice: (Device) -> Unit,
-    onRefresh: () -> Unit,
-) {
-    Surface(
-        color = MaterialTheme.colors.background,
-        modifier = Modifier.fillMaxWidth().height(40.dp),
-    ) {
-        Box {
-            Row(Modifier.align(Alignment.CenterStart).wrapContentSize()) {
-                DropDownDeviceMenu(
-                    devices = state.devices,
-                    selectedDevice = state.selectedDevice,
-                    onSelectDevice = onSelectDevice,
-                    modifier = Modifier.wrapContentWidth(),
-                )
-            }
-
-            Row(
-                horizontalArrangement = Arrangement.spacedBy(4.dp),
-                modifier = Modifier.align(Alignment.CenterEnd).wrapContentSize().padding(4.dp)
-            ) {
-                CommandIconButton(
-                    image = Lucide.Power,
-                    onClick = {},
-                    padding = 2.dp
-                )
-
-                CommandIconButton(
-                    image = Lucide.PowerOff,
-                    onClick = {},
-                    padding = 2.dp
-                )
-
-                CommandIconButton(
-                    image = Lucide.Volume2,
-                    onClick = {}
-                )
-
-                CommandIconButton(
-                    image = Lucide.Volume1,
-                    onClick = {}
-                )
-
-                CommandIconButton(
-                    image = Lucide.Camera,
-                    onClick = {},
-                    padding = 2.dp
-                )
-
-                CommandIconButton(
-                    image = Lucide.Triangle,
-                    degrees = -90f,
-                    onClick = {},
-                    padding = 2.dp
-                )
-
-                CommandIconButton(
-                    image = Lucide.Circle,
-                    onClick = {},
-                    padding = 2.dp
-                )
-
-                CommandIconButton(
-                    image = Lucide.Square,
-                    onClick = {},
-                    padding = 2.dp
-                )
-
-                CommandIconDivider()
-
-                var isPress: Boolean by remember { mutableStateOf(false) }
-                val degrees: Float by animateFloatAsState(if (isPress) -90f else 0f)
-                CommandIconButton(
-                    modifier = Modifier
-                        .onPointerEvent(PointerEventType.Press) { isPress = true }
-                        .onPointerEvent(PointerEventType.Release) { isPress = false },
-                    image = Lucide.RefreshCcw,
-                    degrees = degrees,
-                    onClick = { onRefresh() },
-                    padding = 2.dp
-                )
-            }
         }
     }
 }
